Create 7-Reverse-Integer-2.py
A second solution to the following problem:
https://leetcode.com/problems/reverse-integer/

If it weren't for the condition that the environment does not allow you to store 64-bit integers, the simple solution to the problem would simply to be to convert the integer to a string and reverse the characters. However, in certain situations this would cause a 64-bit integer to be stored. Therefore, I thought it would be fun to figure out a way to implement this solution.