Create Roman-To-Integer.py
Python solution to the following LeetCode problem:
https://leetcode.com/problems/roman-to-integer/

The task was to read in a Roman numeral value and convert it into its corresponding integer value. This version of the solution used subtraction for the cases where a smaller numeral would appear before a larger one to signify a value between them (IV representing 4).