Create 13-Roman-To-Integer-2.py
Second Python solution to the following LeetCode problem:
https://leetcode.com/problems/roman-to-integer/

The task was to read in a Roman numeral value and convert it into its corresponding integer value. 

This version of the solution replaced subtraction cases to a more easily readable string before conversion, so no extraneous calculations needed to be performed.