Create 17-Letter-Combinations-Of-A-Phone-Number.py
Solution to the following LeetCode problem:
https://leetcode.com/problems/letter-combinations-of-a-phone-number/description/

Where a string of digits ranging from 2-9 (inclusive) is given, with each number corresponding to digits on a phone keypad. Given this string of digits, a list containing every possible letter combination for the numbers given was requested. 

This was done by using a recursive method backtrack and explore every possible combination before appending it to the list.