Create 25-Reverse-Nodes-In-K-Group.py
Python solution to the following LeetCode problem: 
https://leetcode.com/problems/remove-duplicates-from-sorted-array/description/

Where we are given an integer array nums sorted in non-decreasing order, remove the duplicates in-place such that each unique element appears only once. The relative order of the elements should be kept the same. Then return the number of unique elements in nums.