Create 26-Remove-Duplicates-From-Sorted-Array.py
Python solution to the following LeetCode problem: 
https://leetcode.com/problems/remove-duplicates-from-sorted-array/description/

Where, given an integer array nums sorted in non-decreasing order, we should remove the duplicates in-place such that each unique element appears only once. The relative order of the elements should be kept the same. Then return the number of unique elements in nums.