Create 0033-Search-In-Rotated-Sorted-Array.py
Python solution to the following LeetCode problem: https://leetcode.com/problems/search-in-rotated-sorted-array/description/

Where we were given an array, nums, after the possible rotation and an integer target, with the intent of returning the index of target if it is in nums, or -1 if it is not in nums. The algorithm must have a O(log n) runtime complexity.