Create 0035-Search-Insert-Position.py
Python solution to the following LeetCode problem: https://leetcode.com/problems/search-insert-position/description/

Where we were given a sorted array of distinct integers and a target value, with the intention of returning the index if the target is found. If not, the index where it would be if it were inserted in order should have been returned. The algorithm was to be written with O(log n) runtime complexity.